backend/data_processor/requirements.txt
asyncio
sqlalchemy
psycopg2-binary
influxdb
python-dotenv==1.0.1